Fixes in response to code review.
* Removes emojis from output
* Auto-generate certificates on the fly
* Move helper methods out of vault_test.go and into helper files.
