+++ tests/certification/secretstores/hashicorp/vault/components/caFamily/Makefile
@@ -8,26 +8,30 @@ all: $(TARGETS)
 # "x509: certificate relies on legacy Common Name field" errors
 
 certificates/key.pem certificates/cert.pem:
+	set -e; \
+	mkdir -v -p certificates; \
 	openssl req -x509 -newkey rsa:4096 \
 		-keyout certificates/key.pem \
 		-out certificates/cert.pem \
 		-sha256 -days 3650 \
 		-nodes \
 		-addext "subjectAltName = DNS:hashicorp_vault,DNS:localhost,IP:127.0.0.1" \
-		-subj "/C=CA/ST=BC/L=Vancouver/O=Dapr Testing/OU=Org/CN=www.dapr.io"
+		-subj "/C=CA/ST=BC/L=Vancouver/O=Dapr Testing/OU=Org/CN=www.dapr.io"; \
 	chmod -v 644 certificates/key.pem certificates/cert.pem
 
 # We use this for negative tests that ensure we reject connecting to
 # a server using a distinct Certificate Authority -- despite the server certificate
 # having all the right identifiers
 certificates/incorrect-ca-key.pem certificates/incorrect-ca-cert.pem:
+	set -e; \
+	mkdir -v -p certificates; \
 	openssl req -x509 -newkey rsa:4096 \
 		-keyout certificates/incorrect-ca-key.pem \
 		-out certificates/incorrect-ca-cert.pem \
 		-sha256 -days 3650 \
 		-nodes \
 		-addext "subjectAltName = DNS:hashicorp_vault,DNS:localhost,IP:127.0.0.1" \
-		-subj "/C=CA/ST=BC/L=Vancouver/O=Dapr Testing/OU=Org/CN=www.dapr.io"
+		-subj "/C=CA/ST=BC/L=Vancouver/O=Dapr Testing/OU=Org/CN=www.dapr.io" ; \
 	chmod -v 644 certificates/incorrect-ca-key.pem certificates/incorrect-ca-cert.pem
 		
 
@@ -43,6 +47,7 @@ caPem/hashicorp-vault.yml: caPem/hashicorp-vault.yml.template certificates/cert.
 
 clean:
 	rm -f -v $(TARGETS)
+	rmdir certificates
 
 .PHONY: clean
 .PHONY: all
