--- tests/certification/secretstores/hashicorp/vault/components/caFamily/caPem/hashicorp-vault.yml
@@ -1,59 +1,0 @@
-#
-# THIS FILE IS AUTO-GENERATED - DO NOT EDIT
-#
-
-
-apiVersion: dapr.io/v1alpha1
-kind: Component
-metadata:
-  name: my-hashicorp-vault-TestCaFamilyOfFields-caPem
-  namespace: default
-spec:
-  type: secretstores.hashicorp.vault
-  version: v1
-  metadata:
-  - name: vaultAddr
-    value: "https://127.0.0.1:8200"
-  # Enforce TLS verification because this test is all about this
-  - name: skipVerify
-    value: false
-  - name: vaultToken  # Matches docker compose VAULT_DEV_ROOT_TOKEN_ID env. var.
-    value: "vault-dev-root-token-id"
-  - name: tlsServerName
-    value: hashicorp_vault
-  - name: caPem
-    value: |-
-          -----BEGIN CERTIFICATE-----
-          MIIF4TCCA8mgAwIBAgIUQz0IMcZaa13YeKVjvz8ESiNCEJgwDQYJKoZIhvcNAQEL
-          BQAwaTELMAkGA1UEBhMCQ0ExCzAJBgNVBAgMAkJDMRIwEAYDVQQHDAlWYW5jb3V2
-          ZXIxFTATBgNVBAoMDERhcHIgVGVzdGluZzEMMAoGA1UECwwDT3JnMRQwEgYDVQQD
-          DAt3d3cuZGFwci5pbzAeFw0yMjExMjUwNDA4NTBaFw0zMjExMjIwNDA4NTBaMGkx
-          CzAJBgNVBAYTAkNBMQswCQYDVQQIDAJCQzESMBAGA1UEBwwJVmFuY291dmVyMRUw
-          EwYDVQQKDAxEYXByIFRlc3RpbmcxDDAKBgNVBAsMA09yZzEUMBIGA1UEAwwLd3d3
-          LmRhcHIuaW8wggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIKAoICAQDQxCUMQAfW
-          RDvmpCYV/2W5hhWutQlPL48oajtxmX8qeiPoiSKbbirOXUnKBqLM2QVRupkjC2qG
-          w4JzbHfxxipaTvpof5Zh53GwomBeeDpAQGWiFpd6ObUoLzxWlN2RuDCZx44fob7g
-          cbmXrMawL6H7qpo92cgMDudh+30ICwNe2lSvbq5p4kMWjxwFSyErvwfAKXKuMtIw
-          5160nLWpbMFlKTj1oYeQWbjDNA168SUGkkaRIC/2uxZ5I9GIQtk8h5NXI0RxDlks
-          GijEVyMtbi25tpefNCX1F1oG/sjLthdliXfSbGLNGKdgiNdvuhJIq0x2aFiNwcrd
-          pPnkweg7k6Ew4v9kewaIh7UttMO9MMGcAsDN+ympYaodgrk/qw4285Bl328sUC21
-          XadkKErcxnEW8bc5JHZkflH/jC5Z+tWThpgQKz6aIw/h9ApXUMpwqcJKrES4+Wyd
-          zZQQeORliiefBcMv6Yq7oE9mzq1vjWyCYqt/dAAOJ1mZGnzGAnKeubtPeAICfbCZ
-          zReChaiyqcM8rvtxpE2ldskroUT312bTIdpP90QVgbGGseJ///y5SjmnAGLEJiiG
-          roVdmA/tmi7buXR5SqOrqL4037rCNY+8QThmZuH24UZCoPPRAbkQp1wF2qQCcVxT
-          IM2tNdR6ItHldE+5TS7uhQ4eJI9XfXs7CwIDAQABo4GAMH4wHQYDVR0OBBYEFH/Z
-          KI7WpyDBM3UZE4sFYp/Rt6WuMB8GA1UdIwQYMBaAFH/ZKI7WpyDBM3UZE4sFYp/R
-          t6WuMA8GA1UdEwEB/wQFMAMBAf8wKwYDVR0RBCQwIoIPaGFzaGljb3JwX3ZhdWx0
-          gglsb2NhbGhvc3SHBH8AAAEwDQYJKoZIhvcNAQELBQADggIBAKvq+GzTZVePm3Cg
-          TRgT9tL0IojzBwEVulQYM1zoBPS9d/djwIznfA5IMJDwZF50uhCOwZqmHUn4GJPB
-          ClkYmxjQlcwfYJjiU4sVPu7zNI/bfm7wZnJLbyUNYnshuBRQ6eianwqEo+ZeZQQd
-          4fM9QyfYyvW2YanPSQLOtvE6nXZzkzpVMB1H+wp6klAQ6sThHqGVpGTeNxUBaVeV
-          dN9vBzIdRSJ0PlTmjCiBykB2HqLEMg9ZXKBHXgciI6APiSd6bd40KNKy8eoayFPQ
-          rx4tdxd8kbIMRyJG4puSEKAVDaUcFqgNXOFctS3d7JX9EGKGm+aWvd8CTEOtAay9
-          VMb1f6mFmMPuGOrK4JQBVOriLVBpHT9QlOIWlj9CuaycHiPfU9p3A1qL72etFdNA
-          1pQ44QYjiBVxpyQvLO1DGm/HrnShWPoyBRaT2lQbMUudqeQMkQQQXRDBK4L3CIwP
-          pBNg/S5rJAdyAZrpCtjK896bGscCplKI6/Io0g2vOcdLrtwAn23ocshL6PWndue8
-          k6XUS1LFLKz/cgRzzk93yD5VMYkYwo0hfHdpsF9AELl6DHsILEvkAYmR8zJcz/o8
-          hpsTwp3eNexA//gED4f4qWoTbnr1ofFToloQqC/nWhtNaO0BNClyJH8cYcVNQLFX
-          LFXkjVxkrXK8v2vNPR2KvyXleKFk
-          -----END CERTIFICATE-----
