--- tests/certification/secretstores/hashicorp/vault/components/caFamily/certificates/incorrect-ca-cert.pem
@@ -1,34 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIF4TCCA8mgAwIBAgIUSrQ0syOfHO5v4Jl7gqdachXf1o8wDQYJKoZIhvcNAQEL
-BQAwaTELMAkGA1UEBhMCQ0ExCzAJBgNVBAgMAkJDMRIwEAYDVQQHDAlWYW5jb3V2
-ZXIxFTATBgNVBAoMDERhcHIgVGVzdGluZzEMMAoGA1UECwwDT3JnMRQwEgYDVQQD
-DAt3d3cuZGFwci5pbzAeFw0yMjExMjUxOTQ0MDJaFw0zMjExMjIxOTQ0MDJaMGkx
-CzAJBgNVBAYTAkNBMQswCQYDVQQIDAJCQzESMBAGA1UEBwwJVmFuY291dmVyMRUw
-EwYDVQQKDAxEYXByIFRlc3RpbmcxDDAKBgNVBAsMA09yZzEUMBIGA1UEAwwLd3d3
-LmRhcHIuaW8wggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIKAoICAQDCdxTPbiMc
-Gx7l/6cBI8nuvrX0AA+OZNoxNRSa/Y5LNXOdIFbMvBHgK4spQKN2RXVjLB9PZ6Tz
-rNxlQNgMlF3dOVGxFVBLhjt1acDPXjeDYeCrq6lj5NnqBUJGE7Weau8KIAh1iDfp
-pFT5Sa7spVwdFpVcgS/WUvB7LMIqZJgCvydMw1DMM8iE7KbuHBeGO0dkELDt1b28
-sYAB2LV1pesr4yIVeBcqDmZJhHhKyXeNi0RxYwVR+c5/AIcy4oz3AwGaj00xCSRS
-tbUZYT3wtKuFWajLVaFhhWspqnEgDfGaYgRWMBhybLUBJSkTi1y6XfGPbJg5p9cr
-rXuLNeqwavJqi3/ow7N6ybCC+gpZFGtotUIVe6G9Gw4puPtpL8dXRt5UGI77lUMA
-cjaRTql4RZG9yiWxvkppbJohYzZj39buVVtnfK4h6fVeOsIrotjq6A07PmrC9IZG
-tYyShDc0lk0ns08PuaFxUEWDFSNgPXzZh7jVXykBRqWkqsAVUhve8NYBB0GyEIJ3
-fl/Nbb0ODqQXrknvD2af9jRrd0yC3NSg3+3pW9bMLe1qA28GiisGREhIV2I7WIm8
-nDQX/sxE1WOVlL5QQP5mnkZEJJIOQoq5AT+orFquvbCUYhyqf5VVM07+H3A/C6Au
-AZ0EBBzBKdF15UvPA68Nqef/OX4jCs584wIDAQABo4GAMH4wHQYDVR0OBBYEFHUA
-p2sagqDSpNEfGHXLQTQPCsC6MB8GA1UdIwQYMBaAFHUAp2sagqDSpNEfGHXLQTQP
-CsC6MA8GA1UdEwEB/wQFMAMBAf8wKwYDVR0RBCQwIoIPaGFzaGljb3JwX3ZhdWx0
-gglsb2NhbGhvc3SHBH8AAAEwDQYJKoZIhvcNAQELBQADggIBAG/8gN+7C+ROZrDS
-fWHYA68jwdQWqe5+5SlZ4sMEbboA7+++csptBi27oQEtCdsTMx/4eC8/Topsz1//
-OcHPo+/pQQW5VWo1Ny0Nfq9TJ+CZ9NuEoizSnLvVWhEBclA7bhkGvp3cqXTa6XqH
-EDSur0ltQRhs0F+zFRPuPqEIJziLv54BEnt1iM2ydv3yVi0dF9Xz8BE92pB93ONd
-PnnvCjFhvZn9X2oogBgf0HaMTQDj02mm25q86NVZ5qSEBuc+NI/8l7s5ruo+HXcY
-q2WqAgwnQfmYUrNqHNmB0zj0+NygZytWmPYb1GVSi+3PcgVKW08EsQ7i8MI00CCY
-/J2Pb/wHcu1VSRjX3ijhcp4MDfHQmFAnRyaey/ocHdUrBiTtWzNszD5DQhMPOKGx
-ksItrRDExgDkVZspqli72oFnZyF3V4NEAo/ZDk2k5o91huyc7d/sQ4mc2ghJfNCT
-aMWsHbYF7oUkwvun4K8duePB0M1JZS4qXG0kmQrEfUSocS3U8AWItvVQ9H0DCbpT
-pcQFBqCuaj3ehuzsExbKLKI6+VrqTXZi8ZPozyxbkWuXnXSPFdG5wn5eQrFQxGnt
-jFjrzTFeo9kzp3a7fYVveR70xYF3hdtHbPRhxt18k8BZB1vLrb08v9df0LAFfcd2
-vljYnSZYxz/14hzgCxFSQq6D92/E
------END CERTIFICATE-----
