--- tests/certification/secretstores/hashicorp/vault/components/caFamily/certificates/incorrect-ca-key.pem
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJRAIBADANBgkqhkiG9w0BAQEFAASCCS4wggkqAgEAAoICAQDCdxTPbiMcGx7l
-/6cBI8nuvrX0AA+OZNoxNRSa/Y5LNXOdIFbMvBHgK4spQKN2RXVjLB9PZ6TzrNxl
-QNgMlF3dOVGxFVBLhjt1acDPXjeDYeCrq6lj5NnqBUJGE7Weau8KIAh1iDfppFT5
-Sa7spVwdFpVcgS/WUvB7LMIqZJgCvydMw1DMM8iE7KbuHBeGO0dkELDt1b28sYAB
-2LV1pesr4yIVeBcqDmZJhHhKyXeNi0RxYwVR+c5/AIcy4oz3AwGaj00xCSRStbUZ
-YT3wtKuFWajLVaFhhWspqnEgDfGaYgRWMBhybLUBJSkTi1y6XfGPbJg5p9crrXuL
-NeqwavJqi3/ow7N6ybCC+gpZFGtotUIVe6G9Gw4puPtpL8dXRt5UGI77lUMAcjaR
-Tql4RZG9yiWxvkppbJohYzZj39buVVtnfK4h6fVeOsIrotjq6A07PmrC9IZGtYyS
-hDc0lk0ns08PuaFxUEWDFSNgPXzZh7jVXykBRqWkqsAVUhve8NYBB0GyEIJ3fl/N
-bb0ODqQXrknvD2af9jRrd0yC3NSg3+3pW9bMLe1qA28GiisGREhIV2I7WIm8nDQX
-/sxE1WOVlL5QQP5mnkZEJJIOQoq5AT+orFquvbCUYhyqf5VVM07+H3A/C6AuAZ0E
-BBzBKdF15UvPA68Nqef/OX4jCs584wIDAQABAoICAQCm8+dY4ryf36TaX284qkou
-f2rlLbwtSuaMgn7LRc6JQT2d82qx4PET3S5fX1A3NRC/8Nx8poXWclzRTSdvPV2M
-lS5Ocvb5sCeR+RYAsmnEkR57JCiMS6jOom8wjLFLbZEFx1Sb2zPtkhkl4J2eK6V2
-j7HxtjOktJYjmdyk2rIrAYOzAqh2eCq5/BagrFuB8ZisDyxnNC7uNhGpBC1d8PGD
-P0A+4Fg60hv9GA7SXj8WtS8wjYu+wlIHdSOdiyTs2Jbi5A6of6nAnKH7RSIRansr
-RDuHPmALwqirH9p5Y58+JNKGUEpcynlnn0ixr98PXaIKuXfDYBZ6ODmt3b88cMTn
-NgQ0AF9SLOqtong3roxq/fluHmD7TbX3sryBmjEUtdKJ4SwIYRbqJt6yKTn99VBJ
-tSBxokz2s8/ExnnYmiOEMmf6Ri8pmszu5hy8tswvVaHY9Vgu+KhF4spfyy6tWUly
-ducfmvnNqpp9f+KID73pIN0OV0Wb/2smvmBeM8AQqh7KOcBvE9Uep7sNZs/6MnuK
-ZodQf4PlwiE50Z05qiYqivy74U/57Sw30nBd0m4B7gWwDKRd0dMxanv2JVQ+AEq7
-Fsue7/CqQbejG5ExspOFb5sFrQLD3MD5//TIAQCjUjEGGMEX/Su+if7viw8b2Ztg
-v2yaEqLYf/nKxyjMQ3zyKQKCAQEA73YxAFgI8/hqmAwvorIvVhVAzPFRq9lYU5ol
-ic8GbQjv403Qlc9mIv5VJulNHIP3jBiDbI0BGyavvZ4emwxd5F+YzUMCfN+M6brL
-opL6/SeBd6clGqYVIYRPMEh5i4hXgfDHZtLEkUhoNULyo+1TtLFiLTufeQl1i3u3
-VeCLhadpg3VQMzo2ADzDDWV4gm+Zj6vdhk4ia20JZZs6+jbehi62JVcOW7wA7A1/
-UotYLNzWJ+WMfG5chLqC4FGA7p0mAz7XBXO+4TNGnoeliBohba5OIvDX2HAiOqk7
-IUzmBzq3ycTn0yQzrk51/1s99TNJ6YtFiLISmrxtpy9K5vcS7QKCAQEAz+VT4ZJV
-B0JSvAuJs/OOTrdaHRG3U/39ZRZ7B/LstUYk0Mw6gNvCB/I66LFjwnv9sv8vwPDy
-32fWyjt4c1mOkAWw8kSi2Z7e+/Tbkuhi2UwrIuF2Zk6nUyP0oDD03xYOryEcgxWd
-JtW4eOA96BU42tApyj+9D3JJshinVD+QgTg6GlxPWqwHQAqMPqOpuEbMJBQhDlZu
-dFtlYv+866hcoeJd54SZhXQqqMD/XOEL+n5CcSGOO0pt5vCAFDoT7Vws/xQ5frx/
-m1LSVu8nbLVXS37NbRRHeKV7aK3jQ6duG7iC1xccWlUKh16CQKgqWz7WEGS1b5dR
-MbYxBd0Zd1TFDwKCAQEArLeNJHhbDM2y5LR1b14iMm7l7Gxqwg7kXHikdr+yHNcD
-X+nydvu+6aKUfJpuGhE9bfkIYjVaBOFkhNoVrTPATUZ+ih65HqSHmNY4pqxpCEuE
-YG9Ynd4JA8HQ9DxyHqhapcirqoh/a2LUdwYwyPgXFg+Thpior76Y3+B+Hhz1xZM4
-0a6+CBUGeilZbsIvp7UjB7Oo9oRI4z8kNfs/ENaerh3nFwuxuYzjBV++5CTJlqI6
-fVBegil/5eQ5Hd7rpjAHCPQ8VxxjIz++wf2I8aAXWGsOxLj+lohayADkO2Khz7kU
-cx3pMCu+cwz4oxSGneM+fosLvkrijGU4mgvMdMblCQKCAQB62xID5p766DXOTJ82
-ycKaP/4v2Y1WRG2fdHaGTweVZHtTLK2SAswaOAqaAmJqxy04v1cwFPPiFBI5doWK
-3SzDbamQsgDuy9UIXj34Bf6MXVfMiCKY+PgmyMPCesvqnbVVxyJ7UKWLEpeuTw4m
-fYwYuxVHIP2QCEEGz0zrG5lA08J6W3zFn421ScZxKfSdIrF2zLJwL6Gcf3pd07WI
-Kfl77TMDIzg6RlkDWvJNp+yVPlHzWxobbbOe2xdFXMDz7xnkL8Z+X9SlLH+Cwlex
-BUUuJyQsxuQpptlUFhaMn4ARP6uKgmzN8R2Eu9QLc5V8UJjeHTWfWXOxBfLYfT2U
-aHL/AoIBAQCPc8rfH3M0ZWv8edsibvOgUtJ3yG4+IiRjLEnNhK6NVZHfX9RIAmqO
-csIYRhV8hjBVLwAc12a3t105JXfbYnQ9ucwsXmSKt0W+CDOCOMKCEvUv/33ePhZN
-dwmyb+E9C32mb/xWipmt/t8ucNUDkX81U7tauEWjx2tQv5Db5iGmuoIkPAxbeOLc
-JRNTyYzBYiG7j3nKo/6Yg69eUZibh1M0Dg6NIZ9ltmU50yB7JQBCxrJGz7VcPmJb
-ute7XRuTHKHjESeTHkC5RsSAOc1cLmcu/W3qWDYZaFAo7pX3e3bWqAbatVaqSs6C
-WG6Z5JfRpAOCbhDnmIuc1NxwhVyG8c3q
------END PRIVATE KEY-----
