--- tests/certification/secretstores/hashicorp/vault/components/caFamily/certificates/key.pem
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJQwIBADANBgkqhkiG9w0BAQEFAASCCS0wggkpAgEAAoICAQDQxCUMQAfWRDvm
-pCYV/2W5hhWutQlPL48oajtxmX8qeiPoiSKbbirOXUnKBqLM2QVRupkjC2qGw4Jz
-bHfxxipaTvpof5Zh53GwomBeeDpAQGWiFpd6ObUoLzxWlN2RuDCZx44fob7gcbmX
-rMawL6H7qpo92cgMDudh+30ICwNe2lSvbq5p4kMWjxwFSyErvwfAKXKuMtIw5160
-nLWpbMFlKTj1oYeQWbjDNA168SUGkkaRIC/2uxZ5I9GIQtk8h5NXI0RxDlksGijE
-VyMtbi25tpefNCX1F1oG/sjLthdliXfSbGLNGKdgiNdvuhJIq0x2aFiNwcrdpPnk
-weg7k6Ew4v9kewaIh7UttMO9MMGcAsDN+ympYaodgrk/qw4285Bl328sUC21Xadk
-KErcxnEW8bc5JHZkflH/jC5Z+tWThpgQKz6aIw/h9ApXUMpwqcJKrES4+WydzZQQ
-eORliiefBcMv6Yq7oE9mzq1vjWyCYqt/dAAOJ1mZGnzGAnKeubtPeAICfbCZzReC
-haiyqcM8rvtxpE2ldskroUT312bTIdpP90QVgbGGseJ///y5SjmnAGLEJiiGroVd
-mA/tmi7buXR5SqOrqL4037rCNY+8QThmZuH24UZCoPPRAbkQp1wF2qQCcVxTIM2t
-NdR6ItHldE+5TS7uhQ4eJI9XfXs7CwIDAQABAoICAQCThohHp9fh6kYKLmVblidk
-ewRHeVcixhD8FlX7aOUcN0O4x2/VNny5Z+MFCLty1smTf7FHuoGEyhQdTMFV8Rts
-xfLedYrnlI4arCQ9nehMDzXC3QQeMzl5NCgHZxtMJkTx5rrgT4R7RuJH9KhTSNXG
-6kAlBQkAvGd2LimkzWOMW+0MK8LEwEd4YZkDhJvkNhJcEO4he3hdZqLeLv+QwnHH
-8uF1crhBoaWXC+mav13SzvtfAsfiK0MuR8mB6PtqcsQEB8ANxH3vVG0KkikhqYJW
-glmS/17zW5Ua7GEo03ylHAe4gy8ukAV3ijvzkCHriGF4KzSouBGLn1GX/g8GSsTU
-Z8xjkQ7/fvL0BxngglmjQOJtqmhGbX9vBwJB7Lufdo2uv3/1+oHVDRgtGx6IldAz
-ZoiZj1owoGs9wxx6EoDnh4L51u/Kg18yjk7VhOU1TLo+bO97pHcEV3T20Vk5ZdPY
-xJ4YSBtaX9RhvEPDK+gZur2tLO0i4B1j5JrqHg+w9ROj2jeYtDP9Ji7TiwJa6crw
-TzSOZq+5Bbdc3LzAmG/1plLbJquusi0bi+GZoS+Cadpv0pAefGxsa8TW+0fDKfxc
-xkRW50Dh3bSv7VCuEyqgGpU2Kqkz0heYa1xCuzYUDtphelBS2Q92gbS2G3nB6zF8
-r/HDS3fE68SSo1hdJvpTAQKCAQEA9Pv8I8ZDHfSiKPUcOQWDjv+t7XkoWqa71Nkl
-B+exmL3MkmzPV7AwYv8mJwdgsildQl92/JE8khNtetLfbmr+AkSYuHi8H96NFMfM
-feQsv16noC2vjH2wHIsR1YcZlvGHK0tK05s7PZxLVWcoHAqzRw0tV/1VdcHq54t/
-QfrqQFdrQ4G+K/iYyT3MqErWdk566JVQbA2HZxhbau0Dm8V2lx9+5FmWo01GNQ2d
-ZyigSUOWnxYWz4tvA9BQvGZ3NKRp6+vgb6zIcFuSL66bwtc3xi6pJm5yKMQR9TKm
-fypq54wUvkYS098C/MSBDxy/mPlzFW4YZb0//n3D2HglOgRNsQKCAQEA2idAu3GH
-+m6skYX5efTBw3Rz3wv+iopo8K/8ToYETbRdtJV1dLXcR2W4b9ebc+IkIkDOicgO
-4P2eai/CSCpoL+/fI0wd2eyfT6oACQMKoz0lgY39gg5/lnNbjrZuwgH2o+epFi5I
-YhZgaXT6/DhuM8ROx0TViqGyWEWz7hgw+fY3WpnmPQqO/83kxh4akzVJetdZzkLc
-XKJtT0aW8MJ+g4TQUE1j3GzBTN3gPOKVIHbqw4+I39ytwSo5CPO4MGOJRSW7mAOH
-YR8fNIhZuWIiXi0RfrgSFGHwRIPXSZhWCMlwfOSmuuJcvM3qkbw1n+SyZVin3Lm3
-hv9yWrYgch8XewKCAQB9a8RaTFZwL54VeedJChmgTCWvPif6wQHIpqLeuS+jEwh4
-rlw4eGtNQ3M06AIJNEQahuVTyMdJwch66RqaXQxN/PBzANzPNBDmYqO+j+KY1sza
-pIkLVA2lj01CFDh1npGlCJb2XrKliC7JOla+hHvkFErykS//aZfpkvFDxkHHTvoL
-RZEm0c7muEzem/HP/73M9SSHLSX1EfM7erwuM5nTJLzCC1ujnxwibSX4ekQVPhUO
-vTBQcQIU4cFzssG0obfYoJn2km7LlxXwc2QrDCPxwlG26zRmynqnHZc/42d3w6+a
-+iUJc1MENVYddopgttTEU+kJrSnmEF0Dpf7pBOsBAoIBAFWPn6vk01g9oCFTUo2q
-mXR3cBrDMYpUzuPXojWKCpYPb9Er7W1hBoxaMQRso+PHhjSJBnizpmJIFqG/KYv0
-Vfi3M+U8LVkTnUg5fNoL/1gW22b8sbkYxqHTjM1kb/xXmMygR4kTjgAqlpXF88qI
-t3BGRj+Rl89PrAtDC1cO1eaz0vIN+mdmhL9/mYK6u+GQvMkZPXfQcW1Dp2u4W+sB
-3Wzn4mwcblY9sXLcct5uX97jxR/+Y8ipXNyuGsCQFmWkJGqiafktKSPGeQ65i7dR
-7fjR5FK3DyA/aWT18LJEtTmtCBlz9JDHx2elu6J9FusnX3CwXOYNURX5inhZhlA8
-MZECggEBAOo4fYcU+75WtZrzS49EBTYoRRHvihHcZwQ8P/J+pP8yuoD+bGbXyciS
-5OV1yLEoScbzARa9jDvQhYSPyQV6QQcsCT45spIeHURzyYY30WhcluvE0U2BC+Sh
-GslAiwA1IcDZ45hdbj9J33iZMzIroG1evnpOngUd6SwfT8da+MfhPNDnwcgy+HUG
-bjOy6/1ds9saFQkKGgmUAt51hiHreCVMUbu+lNaqJRYMIPmx/zgqXhyhqMC+hh1S
-GP/XqE9ylx3K769VsXS22IuhXJC5EaiSzqzyjVP4kzmCqccKBsngqKFGIwuFI25p
-JpgsDDiLHGcoVG2ei6PE1GFlSixtFvI=
------END PRIVATE KEY-----
