+++ tests/certification/secretstores/hashicorp/vault/components/caFamily/vaultConfig/vault_server.hcl
@@ -4,12 +4,3 @@ listener "tcp" {
   tls_cert_file = "/certificates/cert.pem"
   tls_key_file  = "/certificates/key.pem"
 }
-
-api_addr = "https://127.0.0.1:8201"
-
-#backend "file" {
-#  path = "/vault/file"
-#}
-
-#default_lease_ttl = "168h"
-#max_lease_ttl = "720h"
